Stop recalculating the data directory
This code was only meant to be called once, but after the damage the
previous maintainer inflicted on the codebase, the method has ended
up being called repeatedly (flooding logcat with spurious and annoying
messages in the process). Eventually, as the damage is undone, this code
will go back to being called once. In the meantime, I need to be able
to actually read the output from logcat.