Implement login/logout
This implements the main logic for profile-aware logging in and
logging out. The full infrastructure is in place such as contacting
the remote server, saving the resulting credentials if logging in
succeeds, deleting books from a specific account on logging out, and
removing the credentials. Currently, it is only possible to log in
via the Settings area as the catalog has not yet been fully updated.
The settings area is fully functional, however.

Affects https://github.com/io7m/android/issues/5