Attempt to make animations consistent
This sets a default "no animation" for activity transitions in the
theme and then overrides when the final activity closes. It also hijacks
every entry and exit point to override the pending transitions.

See: https://stackoverflow.com/questions/23095840/change-activitys-default-exit-animation
See: https://stackoverflow.com/questions/3389501/activity-transition-in-android
See: https://stackoverflow.com/questions/4633543/overridependingtransition-does-not-work-when-flag-activity-reorder-to-front-is-u

This is better than having tons of "overridePendingTransition" calls
everywhere in the code. This seems to work somewhat more consistently
than the existing setup.

This introduces a SimplifiedActivity class from which *all*
activities in the project *must* inherit, and renames the existing
SimplifiedActivity class to NavigationDrawerActivity, as that's what
it actually did in practice.

Affects https://github.com/io7m/android/issues/16
Fix https://github.com/io7m/android/issues/25