Improve test suite
This adds more tests for the book controller and adds an EventAssertions
class for reducing the size of tests that want to check the properties
of raised events. It also fixes a minor bug with the sync task: Book
status values were not being published for books that were newly received
when syncing.

This also adds a new default method to the book registry that raises an
exception if a requested book does not exist. This eases unit testing.