Refactor the navigation drawer
This was necessary because there was an unpleasant bug hidden in the
code that would take the user back to the profile selection screen when
he/she attempted to access the Manage Accounts option. The type-unsafe
code was replaced with a much cleaner approach where the drawer only
contains values of type NavigationDrawerItemType. Subclasses of the
type are responsible for performing actions when items are clicked,
and for configuring their own item view in order to display icons
and text and so on.

The SimplifiedPart type was removed as it was a good idea when the
navigation drawer only contained five static items, but actively
harmful now that the drawer is populated with an indefinite number
of items with wildly different behaviour.

This refactoring also drastically improves the animation behaviour
when switching between activities, as it waits for the drawer to
close before performing any action; it seems as though the navigation
drawer may have been responsible for some of the issues raised in
bug https://github.com/io7m/android/issues/25.

It also fixes the problem of not being able to open the navigation
drawer on the accounts setting screen. This was simply an oversight
on the part of the previous maintainer: Any activity that subclasses
NavigationDrawerActivity *must* add views to the content frame via
getContentFrame, it's not possible to set a content view directly
via setContentView because this then removes the layout that the
navigation drawer uses in order to actually be displayed!