Initial attempt at using a consistent theme everywhere
The existing code was attempting to look like a Material-themed
app when running on v21 and above, and a Holo-themed app when
running on anything else. This lead to complexity in the code,
a theme-related bug (#26), and a lot of duplicated resource files
that only differed by a few values.

This change removes all of the theme related code and tries to impose
something that looks as similar as we can get to the Material style
without requiring v21 or introducing any of the hellish support
library classes.

This corrects the magenta text issue in #26, and it adds a few
unfortunate but necessary hacks to set a consistent colour scheme
on components that do not have a public API to do so. I've marked
the locations that should be revisited if and when the minimum
required API level of the app changes to 21 with "XXXAPI21"
comments.

Known issue: Activities that would have displayed a "back" arrow in
the top left corner of the action bar now display both a back arrow
and the navigation drawer icon. I do not yet know how to prevent this
from happening. See the MainSettingsAccountActivity class for an
example of this happening.

Fix https://github.com/io7m/android/issues/26