Improve style/theme behaviour
This replaces the CatalogLeftPadded button with a plain
button (as opposed to a padded LinearLayout containing a
TextView). It also removes all explicit use of simplified_button
and simplified_button_text in layouts in favour of moving those to
the Theme. This should ease the future creation of views and layouts
because it's no longer necessary to use those drawables
directly in order to get the correct visual style.

This also replaces any calls to Activity.getApplicationContext()
with use of the Activity itself, as getApplicationContext() is almost
always the wrong thing to use (it will prevent the current theme
being applied to views).

Fix https://github.com/io7m/android/issues/27