Implement bundled credentials and automatic accounts
This adds a principled mechanism for bundling credentials into
the app that does not require those credentials to be checked into
version control.

Create a `credentials.json` file in simplified-app-lfa/credentials.json
with contents similar to:

```
{
  "credentialsByProvider": {
    "urn:0": { "userName": "abcd", "passWord": "1234" },
    "urn:1": { "userName": "efgh", "passWord": "5678" },
    "urn:2": { "userName": "ijkl", "passWord": "9090" }
  }
}
```

The fields of the `credentialsByProvider` structure are consulted
at runtime to try to locate credentials for matching account providers.
In the above example, every time an account is created with provider
`urn:0`, credentials `abcd:1234` will be inserted into the created
account automatically. When an account has bundled credentials, the
user will _never_ be presented with a login screen if those credentials
turn out to be incorrect; incorrect bundled credentials are a bug
and should be treated as such.

Additionally, this adds an `addAutomatically` flag to the
`account_providers.json` file to indicate that accounts should be
instantiated automatically when profiles are opened and/or created.

This also adds a `feature_manage_accounts` flag to optionally
enable/disable the account management settings section.  This also
re-enables the display of multiple accounts in the navigation drawer,
which were previously hidden in order to hide the ability to manage
accounts.

Finally, this introduces the new LFA accounts and adds icons for each.