Initial branding updates
This moves to requiring API level 21 as a minimum, and makes a first
attempt at cleaning up the current theme mess.

This switches the app to using a consistent colour scheme, but does
currently mean that the app is using platform defaults for buttons.

This also introduces the new LFA logo.