Add support for searching bundled collections
This adds support for making use of the recent "search index" feature
added to opdsget. Searching is performed against book titles and uses
the levenshtein distance between search terms and titles to calculate
a score for relevancy. Search terms that are prefixes of words in
book titles are given higher relevancy.